+++ Addon/ElvUI_mMediaTag/modules/misc/instancedifficulty.lua
@@ -74,56 +74,116 @@ local instanceDifficulty = {
 }
 
 local shortNames = {
-	[2451] = "ULD", --Uldaman: Legacy of Tyr
-	[2515] = "AV", --The Azure Vault
-	[2516] = "NO", --The Nokhud Offensive
-	[2519] = "NELT", --Neltharus
-	[2520] = "BH", --Brackenhide Hollow
-	[2521] = "RLP", --Ruby Life Pools
-	[2526] = "AA", --Algeth'ar Academy
-	[2527] = "HOI", --Halls of Infusion
-	[2284] = "SD", --Sanguine Depths
-	[2285] = "SOA", --Spires of Ascension
-	[2286] = "NW", --The Necrotic Wake
-	[2287] = "HA", --Halls of Atonement
-	[2289] = "PF", --Plaguefall
-	[2290] = "MOTS", --Mists of Tirna Scithe
-	[2291] = "DOS", --De Other Side
-	[2293] = "TOP", --Theater of Pain
-	[2441] = "TTVM", --Tazavesh the Veiled Market
+	-- tww
+	[2648] = "ROOK", --The Rookery
+	[2649] = "PSF", --Priory of the Sacred Flame
+	[2651] = "DFC", --Darkflame Cleft
+	[2652] = "SV", --The Stonevault
+	[2660] = "ARAK", --Ara-Kara, City of Echoes
+	[2661] = "BREW", --Cinderbrew Meadery
+	[2662] = "DAWN", --The Dawnbreaker
+	[2669] = "COT", --City of Threads
+	[2773] = "FLOOD", --Operation: Floodgate
+
+	--df
+	[2451]	= "ULD", --Uldaman: Legacy of Tyr
+	[2515]	= "AV", --The Azure Vault
+	[2516]	= "NO", --The Nokhud Offensive
+	[2519]	= "NELTH", --Neltharus
+	[2520]	= "BH", --Brackenhide Hollow
+	[2521]	= "RLP", --Ruby Life Pools
+	[2526]	= "AA", --Algeth'ar Academy
+	[2527]	= "HOI", --Halls of Infusion
+	[2579]	= "DOTI", --Dawn of the Infinite
+
+	-- sl
+	[2284]	= "SD", --Sanguine Depths
+	[2285]	= "SOA", --Spires of Ascension
+	[2286]	= "NW", --The Necrotic Wake
+	[2287]	= "HOA", --Halls of Atonement
+	[2289]	= "PF", --Plaguefall
+	[2290]	= "MISTS", --Mists of Tirna Scithe
+	[2291]	= "DOS", --De Other Side
+	[2293]	= "TOP", --Theater of Pain
+	[2441]	= "TAZ", --Tazavesh, the Veiled Market
+
+	-- bfa
+	[1594]	= "ML", --The MOTHERLODE!!
+	[1754]	= "FH", --Freehold
+	[1762]	= "KR", --Kings' Rest
+	[1763]	= "AD", --Atal'Dazar
+	[1771]	= "DAGOR", --Tol Dagor
+	[1822]	= "SIEGE", --Siege of Boralus
+	[1841]	= "UNDR", --The Underrot
+	[1862]	= "WM", --Waycrest Manor
+	[1864]	= "SOS", --Shrine of the Storm
+	[1877]	= "TOS", --Temple of Sethraliss
+	[2097]	= "MECHA", --Operation: Mechagon
+
+	-- legion
+	[1456]	= "EOA", --Eye of Azshara
+	[1458]	= "NL", --Neltharion's Lair
+	[1466]	= "DT", --Darkheart Thicket
+	[1477]	= "HOV", --Halls of Valor
+	[1492]	= "MOS", --Maw of Souls
+	[1493]	= "VOTW", --Vault of the Wardens
+	[1501]	= "BRH", --Black Rook Hold
+	[1516]	= "ARC", --The Arcway
+	[1544]	= "VH", --Violet Hold
+	[1571]	= "COS", --Court of Stars
+	[1651]	= "KARA", --Return to Karazhan
+	[1677]	= "COEN", --Cathedral of Eternal Night
+	[1753]	= "SOT", --Seat of the Triumvirate
+
+	-- wod
+	[1182]	= "AUCH", --Auchindoun
+	[1175]	= "BSM", --Bloodmaul Slag Mines
+	[1176]	= "SBG", --Shadowmoon Burial Grounds
+	[1195]	= "ID", --Iron Docks
+	[1208]	= "GD", --Grimrail Depot
+	[1209]	= "SR", --Skyreach
+	[1279]	= "EB", --The Everbloom
+	[1358]	= "UBRS", --Upper Blackrock Spire
+
+	-- mop
 	[959] = "SM", --Shado-pan Monastery
 	[960] = "TJS", --Temple of the Jade Serpent
-	[961] = "SB", --Stormstout Brewery
+	[961] = "SSB", --Stormstout Brewery
 	[962] = "GOTSS", --Gate of the Setting Sun
 	[994] = "MSP", --Mogu'Shan Palace
-	[1011] = "SONT", --Siege of Niuzao Temple
-	[1182] = "AU", --Auchindoun
-	[1175] = "BSM", --Bloodmaul Slag Mines
-	[1176] = "SBG", --Shadowmoon Burial Grounds
-	[1195] = "ID", --Iron Docks
-	[1208] = "GD", --Grimrail Depot
-	[1209] = "SR", --Skyreach
-	[1279] = "TE", --The Everbloom
-	[1358] = "UBS", --Upper Blackrock Spire
-	[1456] = "EOA", --Eye of Azshara
-	[1458] = "NL", --Neltharion's Lair
-	[1466] = "DT", --Darkheart Thicket
-	[1477] = "HOV", --Halls of Valor
-	[1492] = "MOS", --Maw of Souls
-	[1493] = "VOW", --Vault of the Wardens
-	[1501] = "BRH", --Black Rook Hold
-	[1516] = "TA", --The Arcway
-	[1544] = "VH", --Violet Hold
-	[1571] = "COS", --Court of Stars
-	[1651] = "RTK", --Return to Karazhan
-	[1677] = "COEN", --Cathedral of Eternal Night
-	[1753] = "SOTT", --Seat of the Triumvirate
-	[1594] = "TM", --The MOTHERLODE!!
-	[2522] = "VOTI", --Vault of the Incarnates
-	[1754] = "FH", -- Freehold
-	[1841] = "UNDR", -- The Underrot
-	[657] = "VP", -- Vortex
-	[2579] = "DOTI", -- Dawn of the Infinite
+	[1011] = "SNT", --Siege of Niuzao Temple
+
+	-- cata
+	[568] = "ZA", --Zul'Aman
+	[643] = "TOTT", --Throne of the Tides
+	[644] = "HOC", --Halls of Origination
+	[645] = "BC", --Blackrock Caverns
+	[657] = "VP", --The Vortex Pinnacle
+	[670] = "GB", --Grim Batol
+	[725] = "SC", --The Stonecore
+	[755] = "LCOTT", --Lost City of the Tol'vir
+	[859] = "ZG", --Zul'Gurub
+	[938] = "ET", --End Time
+	[939] = "WE", --Well of Eternity
+	[940] = "HT", --Hour of Twilight
+
+	--wrath
+	[574]	= "UK", --Utgarde Keep
+	[575]	= "UP", --Utgarde Pinnacle
+	[576]	= "NEXUS", --The Nexus
+	[578]	= "OCULUS", --The Oculus
+	[595]	= "TCOS", --The Culling of Stratholme
+	[599]	= "HS", --Halls of Stone
+	[600]	= "DRAK", --Drak'Tharon Keep
+	[601]	= "NERUB", --Azjol-Nerub
+	[602]	= "HOL", --Halls of Lightning
+	[604]	= "GD", --Gundrak
+	[608]	= "VH", --The Violet Hold
+	[619]	= "AHN", --Ahn'kahet: The Old Kingdom
+	[632]	= "FOS", --The Forge of Souls
+	[650]	= "TOC", --Trial of the Champion
+	[658]	= "POS", --Pit of Saron
+	[668]	= "HOR", --Halls of Reflection
 
 	-- Quest Scenarios
 	[2570] = "TFR",
@@ -202,11 +262,6 @@ function mMT:UpdateColors()
 	instanceDifficulty[216].c = db.quest.color
 	instanceDifficulty[220].c = db.story.color
 
-	-- [205] = { c = "|CFF00FCD2", d = "AI" },
-	-- [208] = { c = "|CFFAB5C07", d = "DELVE" },
-	-- [216] = { c = "|CFFFCA400", d = "QUEST" },
-	-- [220] = { c = "|CFF9700FC", d = "STORY" },
-
 	colors.nhc = db.nhc
 	colors.hc = db.hc
 	colors.m = db.m
@@ -229,20 +284,6 @@ function mMT:UpdateColors()
 	colors.story = db.story
 end
 
-function mMT:ShortName(name)
-	local WordA, WordB, WordC, WordD, WordE, WordF, WordG = strsplit(" ", name, 6)
-
-	WordA = WordA and E:ShortenString(WordA, 1) or ""
-	WordB = WordB and E:ShortenString(WordB, 1) or ""
-	WordC = WordC and E:ShortenString(WordC, 1) or ""
-	WordD = WordD and E:ShortenString(WordD, 1) or ""
-	WordE = WordE and E:ShortenString(WordE, 1) or ""
-	WordF = WordF and E:ShortenString(WordF, 1) or ""
-	WordG = WordG and E:ShortenString(WordG, 1) or ""
-
-	return WordA .. WordB .. WordC .. WordD .. WordE .. WordF .. WordG
-end
-
 local function GetIconSettings(button)
 	local defaults = P.general.minimap.icons[button]
 	local profile = E.db.general.minimap.icons[button]
@@ -257,75 +298,83 @@ local function GetKeystoneLevelandColor()
 		return format("%s%s|r", colors.mpa.color, keyStoneLevel)
 	elseif keyStoneLevel <= 9 then
 		color = mMT:ColorFade(E.db.mMT.instancedifficulty.mpa, E.db.mMT.instancedifficulty.mpb, percentValue[keyStoneLevel])
-		if color then
-			return format("%s%s|r", color.color, keyStoneLevel)
-		end
+		if color then return format("%s%s|r", color.color, keyStoneLevel) end
 	elseif keyStoneLevel <= 14 then
 		color = mMT:ColorFade(E.db.mMT.instancedifficulty.mpb, E.db.mMT.instancedifficulty.mpc, percentValue[keyStoneLevel])
-		if color then
-			return format("%s%s|r", color.color, keyStoneLevel)
-		end
+		if color then return format("%s%s|r", color.color, keyStoneLevel) end
 	elseif keyStoneLevel <= 19 then
 		color = mMT:ColorFade(E.db.mMT.instancedifficulty.mpc, E.db.mMT.instancedifficulty.mpd, percentValue[keyStoneLevel])
-		if color then
-			return format("%s%s|r", color.color, keyStoneLevel)
-		end
+		if color then return format("%s%s|r", color.color, keyStoneLevel) end
 	elseif keyStoneLevel >= 24 then
 		color = mMT:ColorFade(E.db.mMT.instancedifficulty.mpd, E.db.mMT.instancedifficulty.mpe, percentValue[keyStoneLevel])
-		if color then
-			return format("%s%s|r", color.color, keyStoneLevel)
-		end
+		if color then return format("%s%s|r", color.color, keyStoneLevel) end
 	elseif keyStoneLevel >= 29 then
 		color = mMT:ColorFade(E.db.mMT.instancedifficulty.mpe, E.db.mMT.instancedifficulty.mpf, percentValue[keyStoneLevel])
-		if color then
-			return format("%s%s|r", color.color, keyStoneLevel)
-		end
+		if color then return format("%s%s|r", color.color, keyStoneLevel) end
 	else
 		return format("%s%s|r", colors.mpf.color, keyStoneLevel)
 	end
 end
 
-function mMT:GetDungeonInfo(datatext, short, stageBlock, hideDelve)
-	local name, _, difficultyID, _, _, _, _, instanceID, instanceGroupSize, _ = GetInstanceInfo()
-	local _, InstanceType = IsInInstance()
-	name = shortNames[instanceID] or mMT:ShortName(name)
+function mMT:GetCurrentDelveTier()
+	local widgetIDs = {
+		{ id = 6183, tierType = "numbered" },
+		{ id = 6184, tierType = "?" },
+		{ id = 6185, tierType = "??" },
+	}
 
-	if name and not (hideDelve and difficultyID == 208) then
-		local text = ""
-		local difficultyColor = instanceDifficulty[difficultyID] and instanceDifficulty[difficultyID].c or "|CFFFFFFFF"
-		local difficultyShort = instanceDifficulty[difficultyID] and instanceDifficulty[difficultyID].d or ""
-		local isGuildParty = InGuildParty()
-		if E.Retail and difficultyID == 8 and C_MythicPlus.IsMythicPlusActive() and C_ChallengeMode.GetActiveChallengeMapID() then
-			if datatext and not short then
-				text = format("%s%s|r %s%s|r %s", isGuildParty and colors.guild.color or colors.name.color or "|CFFFFFFFF", name, difficultyColor, difficultyShort, GetKeystoneLevelandColor())
-			elseif short then
-				text = format("%s%s|r %s", difficultyColor, difficultyShort, GetKeystoneLevelandColor())
-			else
-				text = format("%s%s|r\n%s%s|r %s", isGuildParty and colors.guild.color or colors.name.color or "|CFFFFFFFF", name, difficultyColor, difficultyShort, GetKeystoneLevelandColor())
-			end
-		elseif InstanceType == "pvp" or InstanceType == "arena" then
-			difficultyColor = instanceDifficulty[34] and instanceDifficulty[34].c or "|CFFFFFFFF"
-			difficultyShort = instanceDifficulty[34] and instanceDifficulty[34].d or ""
-			if datatext and not short then
-				text = format("%s%s|r %s%s|r", isGuildParty and colors.guild.color or colors.name.color or "|CFFFFFFFF", name, difficultyColor, difficultyShort)
-			elseif short then
-				text = format("%s%s|r", difficultyColor, difficultyShort)
-			else
-				text = format("%s%s|r\n%s%s|r", isGuildParty and colors.guild.color or colors.name.color or "|CFFFFFFFF", name, difficultyColor, difficultyShort)
-			end
-		else
-			if datatext and not short then
-				text = format("%s%s|r %s%s|r |CFFF7DC6F%s|r", isGuildParty and colors.guild.color or colors.name.color or "|CFFFFFFFF", name, difficultyColor, difficultyShort, instanceGroupSize)
-			elseif short then
-				text = format("%s%s|r %s%s|r", isGuildParty and colors.guild.color or colors.name.color or "|CFFFFFFFF", name, difficultyColor, difficultyShort)
-			elseif stageBlock then
-				text = format("%s|r %s%s|r", isGuildParty and colors.guild.color or colors.name.color or "|CFFFFFFFF", difficultyColor, difficultyShort)
-			else
-				text = format("%s%s|r\n%s%s|r |CFFF7DC6F%s|r", isGuildParty and colors.guild.color or colors.name.color or "|CFFFFFFFF", name, difficultyColor, difficultyShort, instanceGroupSize)
-			end
+	for _, widget in ipairs(widgetIDs) do
+		local delveInfo = C_UIWidgetManager.GetScenarioHeaderDelvesWidgetVisualizationInfo(widget.id)
+		if delveInfo and delveInfo.shownState == Enum.WidgetShownState.Shown then
+			if widget.tierType ~= "numbered" then return widget.tierType end
+
+			local tierText = delveInfo.tierText or ""
+			return tonumber(tierText:match("%d+")) or 1
 		end
-		return text
 	end
+
+	return 1
+end
+
+local function FormatText(name, difficultyColor, difficultyShort, extraText, isGuildParty)
+    local colorPrefix = isGuildParty and colors.guild.color or colors.name.color or "|CFFFFFFFF"
+    return format("%s%s|r %s%s|r %s", colorPrefix, name, difficultyColor, difficultyShort, extraText or "")
+end
+
+local function GetDifficultyInfo(difficultyID)
+    local difficulty = instanceDifficulty[difficultyID] or {}
+    return difficulty.c or "|CFFFFFFFF", difficulty.d or ""
+end
+
+function mMT:GetDungeonInfo(datatext, short, stageBlock, hideDelve)
+    local name, _, difficultyID, _, _, _, _, instanceID, instanceGroupSize = GetInstanceInfo()
+    local _, instanceType = IsInInstance()
+    name = shortNames[instanceID] or E:ShortenString(name, 5)
+
+    if not name or (hideDelve and difficultyID == 208) then return end
+
+    local difficultyColor, difficultyShort = GetDifficultyInfo(difficultyID)
+    local isGuildParty = InGuildParty()
+    local text = ""
+
+    if E.Retail and difficultyID == 8 and C_MythicPlus.IsMythicPlusActive() and C_ChallengeMode.GetActiveChallengeMapID() then
+        local keystoneText = GetKeystoneLevelandColor()
+        text = short and format("%s%s|r %s", difficultyColor, difficultyShort, keystoneText)
+            or FormatText(name, difficultyColor, difficultyShort, keystoneText, isGuildParty)
+    elseif difficultyID == 208 then
+        local delveTier = mMT:GetCurrentDelveTier()
+        text = format("%s%s|r\n%s%s|r |CFFF7DC6F%s|r", (isGuildParty and colors.guild.color or colors.name.color), name, difficultyColor, difficultyShort, delveTier)
+    elseif instanceType == "pvp" or instanceType == "arena" then
+        difficultyColor, difficultyShort = GetDifficultyInfo(34)
+        text = short and format("%s%s|r", difficultyColor, difficultyShort)
+            or FormatText(name, difficultyColor, difficultyShort, nil, isGuildParty)
+    else
+        local extraText = stageBlock and "" or format("|CFFF7DC6F%s|r", instanceGroupSize)
+        text = short and format("%s%s|r %s%s|r", (isGuildParty and colors.guild.color or colors.name.color), name, difficultyColor, difficultyShort)
+            or FormatText(name, difficultyColor, difficultyShort, extraText, isGuildParty)
+    end
+
+    return text
 end
 
 function UpdateDifficulty()
@@ -333,21 +382,15 @@ function UpdateDifficulty()
 	local difficultyGuild = _G.GuildInstanceDifficulty
 	local battlefieldFrame = _G.MiniMapBattlefieldFrame
 
-	if difficulty then
-		difficulty:Hide()
-	end
-	if difficultyGuild then
-		difficultyGuild:Hide()
-	end
-	if battlefieldFrame then
-		battlefieldFrame:Hide()
-	end
+	if difficulty then difficulty:Hide() end
+	if difficultyGuild then difficultyGuild:Hide() end
+	if battlefieldFrame then battlefieldFrame:Hide() end
 
-	local name, _, _, _, _, _, _, _, _, _ = GetInstanceInfo()
+	local name, _, diffID, _, _, _, _, instanceID, _, _ = GetInstanceInfo()
 	local inInstance, _ = IsInInstance()
 
 	if mIDF then
-		if inInstance and name then
+		if inInstance or (diffID == 208 and C_DelvesUI.HasActiveDelve(instanceID)) and name then
 			mIDF.Text:SetText(mMT:GetDungeonInfo())
 			mIDF:Show()
 		else